adding a little mod with image lines header
the SEQ field is NOT used anymore in ROS , it is still there for
back compatibility. even if you modify this field it will be overwritten
by ROS:

http://answers.ros.org/question/55126/why-does-ros-overwrite-my-sequence-number/

so, do not use it as a reference.

moreover, i modified the MIN DISPARITY field in the launcher with 0, from the
tests with Daniele Cattaneo it works better in this way.